Module audio_streams::async_api

Traits required by audio_streams to perform async operations.

Consumers must provide an implementation of AudioStreamsExecutor to allow the async audio_streams interface to do async io on the Executor provided by the consumer.

These traits follow the interface of cros_async, since it is used by both crosvm and libcras to implement them.

The implementation is provided in cros_async::audio_streams_async.

Traits

AudioStreamsExecutor

Trait of Executor functionality used by audio_streams.

ReadAsync

ReadWriteAsync

WriteAsync

Type Definitions

AsyncStream