Module audio_streams::shm_streams

Structs

MockShmStream

MockShmStreamSource
Source of MockShmStream objects.

NullShmStream
Class that implements ShmStream trait but does nothing with the samples

NullShmStreamSource
Source of NullShmStream objects.

ServerRequest
ServerRequest represents an active request from the server for the client to provide a buffer in shared memory to playback from or capture to.

Enums

Error

Traits

BufferSet
BufferSet is used as a callback mechanism for ServerRequest objects. It is meant to be implemented by the audio stream, allowing arbitrary code to be run after a buffer offset and length is set.

ShmStream
ShmStream allows a client to interact with an active CRAS stream.

ShmStreamSource
ShmStreamSource creates streams for playback or capture of audio.