Trait audio_streams::StreamControl

pub trait StreamControl: Send + Sync {
    fn set_volume(&mut self, _scaler: f64) { ... }

    fn set_mute(&mut self, _mute: bool) { ... }
}

StreamControl provides a way to set the volume and mute states of a stream. StreamControl is separate from the stream so it can be owned by a different thread if needed.

Provided methods

fn set_volume(&mut self, _scaler: f64)

fn set_mute(&mut self, _mute: bool)

Loading content...