Trait audio_streams::StreamSource

pub trait StreamSource: Send {
    fn new_playback_stream(
        &mut self, 
        num_channels: usize, 
        format: SampleFormat, 
        frame_rate: u32, 
        buffer_size: usize
    ) -> Result<(Box<dyn StreamControl>, Box<dyn PlaybackBufferStream>), BoxError>;

    fn new_async_playback_stream(
        &mut self, 
        _num_channels: usize, 
        _format: SampleFormat, 
        _frame_rate: u32, 
        _buffer_size: usize, 
        _ex: &Executor
    ) -> Result<(Box<dyn StreamControl>, Box<dyn AsyncPlaybackBufferStream>), BoxError> { ... }

    fn new_capture_stream(
        &mut self, 
        num_channels: usize, 
        format: SampleFormat, 
        frame_rate: u32, 
        buffer_size: usize, 
        _effects: &[StreamEffect]
    ) -> Result<(Box<dyn StreamControl>, Box<dyn CaptureBufferStream>), BoxError> { ... }

    fn new_async_capture_stream(
        &mut self, 
        num_channels: usize, 
        format: SampleFormat, 
        frame_rate: u32, 
        buffer_size: usize, 
        _effects: &[StreamEffect], 
        _ex: &Executor
    ) -> Result<(Box<dyn StreamControl>, Box<dyn AsyncCaptureBufferStream>), BoxError> { ... }

    fn keep_fds(&self) -> Option<Vec<RawFd>> { ... }
}

StreamSource creates streams for playback or capture of audio.

Required methods

fn new_playback_stream(
&mut self,
num_channels: usize,
format: SampleFormat,
frame_rate: u32,
buffer_size: usize
) -> Result<(Box<dyn StreamControl>, Box<dyn PlaybackBufferStream>), BoxError>

Returns a stream control and buffer generator object. These are separate as the buffer generator might want to be passed to the audio stream.

Provided methods

fn new_async_playback_stream(
&mut self,
_num_channels: usize,
_format: SampleFormat,
_frame_rate: u32,
_buffer_size: usize,
_ex: &Executor
) -> Result<(Box<dyn StreamControl>, Box<dyn AsyncPlaybackBufferStream>), BoxError>

Returns a stream control and async buffer generator object. These are separate as the buffer generator might want to be passed to the audio stream.

fn new_capture_stream(
&mut self,
num_channels: usize,
format: SampleFormat,
frame_rate: u32,
buffer_size: usize,
_effects: &[StreamEffect]
) -> Result<(Box<dyn StreamControl>, Box<dyn CaptureBufferStream>), BoxError>

Returns a stream control and buffer generator object. These are separate as the buffer generator might want to be passed to the audio stream. Default implementation returns NoopStreamControl and NoopCaptureStream.

fn new_async_capture_stream(
&mut self,
num_channels: usize,
format: SampleFormat,
frame_rate: u32,
buffer_size: usize,
_effects: &[StreamEffect],
_ex: &Executor
) -> Result<(Box<dyn StreamControl>, Box<dyn AsyncCaptureBufferStream>), BoxError>

Returns a stream control and async buffer generator object. These are separate as the buffer generator might want to be passed to the audio stream. Default implementation returns NoopStreamControl and NoopCaptureStream.

fn keep_fds(&self) -> Option<Vec<RawFd>>

Returns any open file descriptors needed by the implementor. The FD list helps users of the StreamSource enter Linux jails making sure not to close needed FDs.